Extract normalized grayscale from SmoothGrad
