+++ tf_explain/utils/image.py
@@ -1,5 +1,6 @@
 """ Module for image operations """
 import numpy as np
+import tensorflow as tf
 
 
 def apply_grey_patch(image, top_left_x, top_left_y, patch_size):
@@ -21,3 +22,25 @@ def apply_grey_patch(image, top_left_x, top_left_y, patch_size):
     ] = 127.5
 
     return patched_image
+
+
+@tf.function
+def transform_to_normalized_grayscale(tensor):
+    """
+    Transform tensor over RGB axis to grayscale.
+
+    Args:
+        tensor (tf.Tensor): 4D-Tensor with shape (batch_size, H, W, 3)
+
+    Returns:
+        tf.Tensor: 4D-Tensor of grayscale tensor, with shape (batch_size, H, W, 1)
+    """
+    grayscale_tensor = tf.reduce_sum(tensor, axis=-1)
+    normalized_tensor = tf.cast(
+        255
+        * (grayscale_tensor - tf.reduce_min(grayscale_tensor))
+        / (tf.reduce_max(grayscale_tensor) - tf.reduce_min(grayscale_tensor)),
+        tf.uint8,
+    )
+
+    return normalized_tensor
